requirements.txt
streamlit
praw
sentence-transformers
transformers
faiss-cpu
torch
requests
